Add Xcode support to build_rive.sh
Add an "xcode" option that generates and builds an xcode workspace.

Add -Wshorten-64-to-32 to non-Xcode builds and fix more warnings.

Diffs=
f0da6a7a0 Add Xcode support to build_rive.sh (#7780)

+++ include/rive/pls/pls_draw.hpp
@@ -155,7 +155,7 @@ public:
     pls::ContourDirections m_contourDirections;
 
     // Used to guarantee m_pathRef doesn't change for the entire time we hold it.
-    RIVE_DEBUG_CODE(size_t m_rawPathMutationID;)
+    RIVE_DEBUG_CODE(uint64_t m_rawPathMutationID;)
 };
 
 // Draws a path by fanning tessellation patches around the midpoint of each contour.
